Move BlockMasterClient behind an interface
