+++ core/client/src/main/java/alluxio/client/block/BlockMasterClient.java
@@ -11,64 +11,22 @@
 
 package alluxio.client.block;
 
-import alluxio.AbstractMasterClient;
-import alluxio.Constants;
+import alluxio.Client;
 import alluxio.exception.AlluxioException;
 import alluxio.exception.ConnectionFailedException;
-import alluxio.thrift.AlluxioService;
-import alluxio.thrift.AlluxioTException;
-import alluxio.thrift.BlockMasterClientService;
 import alluxio.wire.BlockInfo;
-import alluxio.wire.ThriftUtils;
 import alluxio.wire.WorkerInfo;
 
-import org.apache.thrift.TException;
-
 import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
- * A wrapper for the thrift client to interact with the block master, used by alluxio clients.
- *
- * Since thrift clients are not thread safe, this class is a wrapper to provide thread safety, and
- * to provide retries.
+ * A client to use for interacting with a Block Master.
  */
 @ThreadSafe
-public final class BlockMasterClient extends AbstractMasterClient {
-  private BlockMasterClientService.Client mClient = null;
-
-  /**
-   * Creates a new block master client.
-   *
-   * @param masterAddress the master address
-   */
-  public BlockMasterClient(InetSocketAddress masterAddress) {
-    super(masterAddress);
-  }
-
-  @Override
-  protected AlluxioService.Client getClient() {
-    return mClient;
-  }
-
-  @Override
-  protected String getServiceName() {
-    return Constants.BLOCK_MASTER_CLIENT_SERVICE_NAME;
-  }
-
-  @Override
-  protected long getServiceVersion() {
-    return Constants.BLOCK_MASTER_CLIENT_SERVICE_VERSION;
-  }
-
-  @Override
-  protected void afterConnect() throws IOException {
-    mClient = new BlockMasterClientService.Client(mProtocol);
-  }
+public interface BlockMasterClient extends Client {
 
   /**
    * Gets the info of a list of workers.
@@ -77,19 +35,8 @@ public final class BlockMasterClient extends AbstractMasterClient {
    * @throws IOException if an I/O error occurs
    * @throws ConnectionFailedException if network connection failed
    */
-  public synchronized List<WorkerInfo> getWorkerInfoList()
-      throws IOException, ConnectionFailedException {
-    return retryRPC(new RpcCallable<List<WorkerInfo>>() {
-      @Override
-      public List<WorkerInfo> call() throws TException {
-        List<WorkerInfo> result = new ArrayList<>();
-        for (alluxio.thrift.WorkerInfo workerInfo : mClient.getWorkerInfoList()) {
-          result.add(ThriftUtils.fromThrift(workerInfo));
-        }
-        return result;
-      }
-    });
-  }
+  List<WorkerInfo> getWorkerInfoList()
+      throws IOException, ConnectionFailedException;
 
   /**
    * Returns the {@link BlockInfo} for a block id.
@@ -99,15 +46,8 @@ public final class BlockMasterClient extends AbstractMasterClient {
    * @throws AlluxioException if an Alluxio error occurs
    * @throws IOException if an I/O error occurs
    */
-  public synchronized BlockInfo getBlockInfo(final long blockId)
-      throws AlluxioException, IOException {
-    return retryRPC(new RpcCallableThrowsAlluxioTException<BlockInfo>() {
-      @Override
-      public BlockInfo call() throws AlluxioTException, TException {
-        return ThriftUtils.fromThrift(mClient.getBlockInfo(blockId));
-      }
-    });
-  }
+  BlockInfo getBlockInfo(final long blockId)
+      throws AlluxioException, IOException;
 
   /**
    * Gets the total Alluxio capacity in bytes, on all the tiers of all the workers.
@@ -116,14 +56,7 @@ public final class BlockMasterClient extends AbstractMasterClient {
    * @throws ConnectionFailedException if network connection failed
    * @throws IOException if an I/O error occurs
    */
-  public synchronized long getCapacityBytes() throws ConnectionFailedException, IOException {
-    return retryRPC(new RpcCallable<Long>() {
-      @Override
-      public Long call() throws TException {
-        return mClient.getCapacityBytes();
-      }
-    });
-  }
+  long getCapacityBytes() throws ConnectionFailedException, IOException;
 
   /**
    * Gets the total amount of used space in bytes, on all the tiers of all the workers.
@@ -132,12 +65,5 @@ public final class BlockMasterClient extends AbstractMasterClient {
    * @throws ConnectionFailedException if network connection failed
    * @throws IOException if an I/O error occurs
    */
-  public synchronized long getUsedBytes() throws ConnectionFailedException, IOException {
-    return retryRPC(new RpcCallable<Long>() {
-      @Override
-      public Long call() throws TException {
-        return mClient.getUsedBytes();
-      }
-    });
-  }
+  long getUsedBytes() throws ConnectionFailedException, IOException;
 }
