+++ core/common/src/main/java/alluxio/AbstractClient.java
@@ -32,7 +32,6 @@ import org.apache.thrift.transport.TTransportException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.io.Closeable;
 import java.io.IOException;
 import java.net.InetSocketAddress;
 
@@ -42,7 +41,7 @@ import javax.annotation.concurrent.ThreadSafe;
  * The base class for clients.
  */
 @ThreadSafe
-public abstract class AbstractClient implements Closeable {
+public abstract class AbstractClient implements Client {
 
   private static final Logger LOG = LoggerFactory.getLogger(Constants.LOGGER_TYPE);
   /** The number of times to retry a particular RPC. */
